Set all time variables as unsigned
Oversight here that would've caused problems ~25 days out.

+++ Arduino/DJLucio/DJLucio.ino
@@ -28,11 +28,11 @@ const int8_t HorizontalSens = 5;  // Mouse sensitivity multipler - 6 max
 const int8_t VerticalSens = 2;    // Mouse sensitivity multipler - 6 max
 
 // Tuning Options
-const uint16_t UpdateRate = 4;  // Controller polling rate, in milliseconds (ms)
-const uint16_t DetectTime = 1000;  // Time before a connected controller is considered stable (ms)
-const uint16_t ConnectRate = 500;  // Rate to attempt reconnections, in ms
-const uint8_t  EffectThreshold = 10;  // Threshold to trigger abilities from the fx dial, 10 = 1/3rd of a revolution
-const uint16_t EffectsTimeout = 1200;  // Timeout for the effects tracker, in ms
+const unsigned long UpdateRate = 4;  // Controller polling rate, in milliseconds (ms)
+const unsigned long DetectTime = 1000;  // Time before a connected controller is considered stable (ms)
+const unsigned long ConnectRate = 500;  // Rate to attempt reconnections, in ms
+const uint8_t       EffectThreshold = 10;  // Threshold to trigger abilities from the fx dial, 10 = 1/3rd of a revolution
+const unsigned long EffectsTimeout = 1200;  // Timeout for the effects tracker, in ms
 
 // Debug Flags (uncomment to add)
 // #define DEBUG                // Enable to use any prints
